Reorganisation, made PropertiesWidget a base class.
Added comments and guidance.
Moved properties_widget into flexlay/gui.
Made comments less supertux biased in flexlay.
